hero-matchup-drafter/requirements.txt
scrapy
pandas
pytest
pyyaml